requirements.txt
flask_bcrypt
pyotp
flask~=2.1.2
flask_socketio
flask_sqlalchemy
SQLAlchemy~=1.4.39
pyjwt